refactor shell func into Helper class
